# Conflicts:
#	Applications/Quickstarts.Servers/ReferenceServer/ReferenceNodeManager.cs
#	Applications/Quickstarts.Servers/ReferenceServer/ReferenceServer.cs
#	Applications/Quickstarts.Servers/TestData/HistoryArchive.cs
#	Applications/Quickstarts.Servers/TestData/HistoryDataReader.cs
#	Applications/Quickstarts.Servers/TestData/HistoryFile.cs
#	Directory.Packages.props
#	Libraries/Opc.Ua.Server/Historian/InMemory/InMemoryHistorianOptions.cs
#	Tests/Opc.Ua.Core.Security.Tests/AuditingOperationTests.cs
#	Tests/Opc.Ua.Core.Security.Tests/CertSessionContext.cs
#	Tests/Opc.Ua.Core.Security.Tests/SecurityCertValidationDepthTests.cs
#	Tests/Opc.Ua.Core.Security.Tests/SecurityCertValidationTests.cs
#	Tests/Opc.Ua.Core.Security.Tests/SecurityCertificateTests.cs
#	Tests/Opc.Ua.Core.Security.Tests/SecurityNoneSession10Tests.cs
#	Tests/Opc.Ua.Core.Security.Tests/SecurityRoleServerBase2Tests.cs
#	Tests/Opc.Ua.Core.Security.Tests/SecurityTests.cs
#	Tests/Opc.Ua.Core.Security.Tests/TestCertificateFactory.cs
#	Tests/Opc.Ua.Core.Tests/Types/BuiltIn/DataValueBenchmarkPayloads.cs
#	Tests/Opc.Ua.Gds.Tests/GdsApplicationDirectoryTests.cs
#	Tests/Opc.Ua.Gds.Tests/GdsDepthTests.cs
#	Tests/Opc.Ua.Gds.Tests/GdsQueryApplicationsTests.cs
#	Tests/Opc.Ua.Gds.Tests/GdsTestFixture.cs
#	Tests/Opc.Ua.History.Tests/AlarmsAndConditionsInstancesTests.cs
#	Tests/Opc.Ua.History.Tests/DataAccessDepthTests.cs
#	Tests/Opc.Ua.History.Tests/FileSystemTests.cs
#	Tests/Opc.Ua.History.Tests/HistoricalAccessDepthTests.cs
#	Tests/Opc.Ua.InformationModel.Tests/AliasnameBaseTests.cs
#	Tests/Opc.Ua.InformationModel.Tests/AttributeReadComplexTests.cs
#	Tests/Opc.Ua.InformationModel.Tests/AttributeWriteIndexTests.cs
#	Tests/Opc.Ua.InformationModel.Tests/AttributeWriteTests.cs
#	Tests/Opc.Ua.InformationModel.Tests/BaseInfoBehavioralTests.cs
#	Tests/Opc.Ua.InformationModel.Tests/BaseInfoParityTests.cs
#	Tests/Opc.Ua.InformationModel.Tests/BaseInformationTests.cs
#	Tests/Opc.Ua.InformationModel.Tests/ViewDepthTests.cs
#	Tests/Opc.Ua.Lds.Tests/DiscoveryDepthTests.cs
#	Tests/Opc.Ua.Lds.Tests/DiscoveryEndpointTests.cs
#	Tests/Opc.Ua.Lds.Tests/DiscoveryFilterTests.cs
#	Tests/Opc.Ua.Lds.Tests/LdsFixtureSmokeTests.cs
#	Tests/Opc.Ua.Lds.Tests/LdsMeConformanceTests.cs
#	Tests/Opc.Ua.Lds.Tests/MiscellaneousTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/MonitorDeadbandFilterTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/MonitorTriggeringTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/MonitoredItemDepthTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/MonitoredItemServicesTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/PublishTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/SubscriptionBasicDepthTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/SubscriptionBasicTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/SubscriptionMinimumTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/SubscriptionMultipleTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/SubscriptionPublishTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/SubscriptionServicesTests.cs
#	Tests/Opc.Ua.Subscriptions.Tests/SubscriptionTransferDepthTests.cs
#	Tests/Opc.Ua.Types.Tests/BuiltIn/DataValueTests.cs

+++ Libraries/Opc.Ua.Client/FileSystem/UaFileInfo.cs
@@ -39,7 +39,7 @@ namespace Opc.Ua.Client.FileSystem
 {
     /// <summary>
     /// Strongly-typed handle to a single OPC UA <c>FileType</c>
-    /// instance. Mirrors <see cref="System.IO.FileInfo"/>.
+    /// instance. Mirrors <see cref="FileInfo"/>.
     /// </summary>
     /// <remarks>
     /// File metadata properties (<see cref="Size"/>,
@@ -111,7 +111,7 @@ namespace Opc.Ua.Client.FileSystem
             Size = metadata.Size ?? 0UL;
             Writable = metadata.Writable ?? false;
             UserWritable = metadata.UserWritable ?? false;
-            OpenCount = metadata.OpenCount ?? (ushort)0;
+            OpenCount = metadata.OpenCount ?? 0;
             MimeType = metadata.MimeType;
             MaxByteStringLength = metadata.MaxByteStringLength;
             LastModifiedTime = metadata.LastModifiedTime;
